+++ package.json
@@ -221,7 +221,7 @@
     "@lobehub/icons": "^4.1.0",
     "@lobehub/market-sdk": "0.29.2",
     "@lobehub/tts": "^4.0.2",
-    "@lobehub/ui": "^4.35.0",
+    "@lobehub/ui": "^4.36.2",
     "@modelcontextprotocol/sdk": "^1.25.3",
     "@napi-rs/canvas": "^0.1.88",
     "@neondatabase/serverless": "^1.0.2",
