+++ scripts/electronWorkflow/modifiers/settingsContentToStatic.mts
@@ -12,13 +12,6 @@ interface DynamicImportInfo {
   start: number;
 }
 
-const isBusinessFeaturesEnabled = () => {
-  const raw = process.env.ENABLE_BUSINESS_FEATURES;
-  if (!raw) return false;
-  const normalized = raw.trim().toLowerCase();
-  return normalized === 'true' || normalized === '1';
-};
-
 const extractDynamicImportsFromMap = (code: string): DynamicImportInfo[] => {
   const results: DynamicImportInfo[] = [];
 
@@ -44,20 +37,12 @@ const extractDynamicImportsFromMap = (code: string): DynamicImportInfo[] => {
   return results;
 };
 
-const generateStaticImports = (imports: DynamicImportInfo[], keepBusinessTabs: boolean): string => {
-  return imports
-    .filter((imp) => keepBusinessTabs || !imp.importPath.includes('@/business/'))
-    .map((imp) => `import ${imp.componentName} from '${imp.importPath}';`)
-    .join('\n');
+const generateStaticImports = (imports: DynamicImportInfo[]): string => {
+  return imports.map((imp) => `import ${imp.componentName} from '${imp.importPath}';`).join('\n');
 };
 
-const generateStaticComponentMap = (
-  imports: DynamicImportInfo[],
-  keepBusinessTabs: boolean,
-): string => {
-  const entries = imports
-    .filter((imp) => keepBusinessTabs || !imp.importPath.includes('@/business/'))
-    .map((imp) => `  [SettingsTabs.${imp.key}]: ${imp.componentName},`);
+const generateStaticComponentMap = (imports: DynamicImportInfo[]): string => {
+  const entries = imports.map((imp) => `  [SettingsTabs.${imp.key}]: ${imp.componentName},`);
 
   return `const componentMap: Record<string, React.ComponentType<{ mobile?: boolean }>> = {\n${entries.join('\n')}\n}`;
 };
@@ -79,13 +64,6 @@ export const convertSettingsContentToStatic = async (TEMP_DIR: string) => {
     filePath,
     name: 'convertSettingsContentToStatic',
     transformer: (code) => {
-      const keepBusinessTabs = isBusinessFeaturesEnabled();
-      if (keepBusinessTabs) {
-        console.log(
-          '    ENABLE_BUSINESS_FEATURES is enabled, preserving business Settings tabs in componentMap',
-        );
-      }
-
       const imports = extractDynamicImportsFromMap(code);
 
       invariant(
@@ -95,8 +73,8 @@ export const convertSettingsContentToStatic = async (TEMP_DIR: string) => {
 
       console.log(`    Found ${imports.length} dynamic imports in componentMap`);
 
-      const staticImports = generateStaticImports(imports, keepBusinessTabs);
-      const staticComponentMap = generateStaticComponentMap(imports, keepBusinessTabs);
+      const staticImports = generateStaticImports(imports);
+      const staticComponentMap = generateStaticComponentMap(imports);
 
       let result = code;
