+++ src/app/[variants]/initialize.ts
